# server/src/openapi/openapi.yaml
...
openapi: 3.0.3
info:
  #  title: Mentor Success API
  description: API for the Mentor Success application
  contact:
    email: whitney@aiontechnology.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.1.0
servers:
  - url: "https://lj6z99n04d.execute-api.us-west-2.amazonaws.com/mentorsuccess-test-service-deployment"
...
paths:
...
  /api/v1/books:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: header
        name: Accept
        schema:
          type:string

    # OPTIONS for /api/v1/books
    options:
      tags:
        - resource
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # POST for /api/v1/books
    post:
      summary: Add a book
      description: Add a book to the system
      tags:
        - resource
      operationId: createBook
      requestBody:
        description: Book to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookModel'
      responses:
        201:
          description: The book was created
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResultModel'
          links:
            self:
              operationId: getBook
              parameters:
                personId: '$response.body#/id'
              description: >
                The `id` value returned in the response can be used as the
                `bookId` parameter in `GET /books/{bookId}`.
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/books"
        responses:
          201:
            statusCode: "201"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "POST"
        type: "http"

    # GET for /api/v1/books
    get:
      summary: Get all books
      description: Get a collection of all books
      tags:
        - resource
      operationId: getAllBooks
      responses:
        200:
          description: A collection of all books
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResultModelCollection'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/books"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.Accept: "method.request.header.Accept"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

  # /api/v1/books/{bookId} ---------------------------------------------------------
  /api/v1/books/{bookId}:
    parameters:
      - in: header
        name: Authorization
        schema:
          type: string
      - in: path
        name: bookId
        schema:
          type: string
          format: uuid
        required: true
        description: The id of the desired book

    # OPTIONS for /api/v1/books/{bookId}
    options:
      tags:
        - resource
      responses:
        200:
          description: Success
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: "string"
            Access-Control-Allow-Methods:
              schema:
                type: "string"
            Access-Control-Allow-Headers:
              schema:
                type: "string"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Empty'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        responses:
          default:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Methods: "'DELETE,GET,HEAD,OPTIONS,PATCH,POST,PUT'"
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization,X-Amz-Date,X-Api-Key,X-Amz-Security-Token'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestTemplates:
          application/json: "{\"statusCode\": 200}"
        passthroughBehavior: "when_no_match"
        type: "mock"

    # GET for /api/v1/books/{bookId}
    get:
      summary: Get a book
      description: Get a book by id
      tags:
        - resource
      operationId: getBook
      responses:
        200:
          description: Found the requested book
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/books/{bookId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.bookId: "method.request.path.bookId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "GET"
        type: "http"

    # PUT for /api/v1/books/{bookId}
    put:
      summary: Update a book
      description: Update the book with the given id
      tags:
        - resource
      operationId: updateBook
      requestBody:
        description: The updated book
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BookModel'
      responses:
        200:
          description: The book was updated
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookResultModel'
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/books/{bookId}"
        responses:
          200:
            statusCode: "200"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.bookId: "method.request.path.bookId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "PUT"
        type: "http"

    # DELETE for /api/v1/books/{bookId}
    delete:
      summary: Disable book
      description: Disable a book record
      tags:
        - resource
      operationId: deactivateBook
      responses:
        204:
          description: The book record was deactivated
          headers:
            Access-Control-Allow-Origin:
              schema:
                type: string
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/NotAuthenticated'
        403:
          $ref: '#/components/responses/NotAuthorized'
        404:
          $ref: '#/components/responses/NotFound'
        500:
          $ref: '#/components/responses/Error'
      x-amazon-apigateway-auth:
        type: None
      x-amazon-apigateway-integration:
        uri: "${lb_url}/api/v1/books/{bookId}"
        responses:
          204:
            statusCode: "204"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          400:
            statusCode: "400"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          401:
            statusCode: "401"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          403:
            statusCode: "403"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          404:
            statusCode: "404"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
          default:
            statusCode: "500"
            responseParameters:
              method.response.header.Access-Control-Allow-Origin: "'*'"
        requestParameters:
          integration.request.header.Authorization: "method.request.header.Authorization"
          integration.request.header.X-Forwarded-Proto: "'https'"
          integration.request.path.bookId: "method.request.path.bookId"
          integration.request.header.X-Forwarded-Host: "stageVariables.api_url"
        passthroughBehavior: "when_no_match"
        connectionType: "VPC_LINK"
        connectionId: "$${stageVariables.vpc_link_id}"
        httpMethod: "DELETE"
        type: "http"

  # /api/v1/games ------------------------------------------------------------------
...
components:
  responses:
    BadRequest:
      description: There was an error while creating the person
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestModel'
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    Error:
      description: Encountered a server error
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthenticated:
      description: Access token is missing or invalid
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotAuthorized:
      description: User does not have permissions for operation
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
    NotFound:
      description: Requested resource could not be found
      headers:
        Access-Control-Allow-Origin:
          schema:
            type: string
...
  schemas:
...
    BadRequestModel:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: string
        error:
          type: object
        message:
          type: string
        path:
          type: string

    # BehaviorModelCollection ------------------------------------------------------
...
    BookModel:
      required:
        - author
        - title
        - gradeLevel
      type: object
      properties:
        author:
          type: string
          maxLength: 40
          example: J. R. R. Tolkien
        title:
          type: string
          maxLength: 30
          example: The Fellowship of the Ring
        gradeLevel:
          type: integer
          example: 3
        interests:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/NameModel'
        leadershipSkills:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/NameModel'
        leaderTraits:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/NameModel'

    # BookResultModel --------------------------------------------------------------
    BookResultModel:
      type: object
      allOf:
        - $ref: '#/components/schemas/BookModel'
      properties:
        _links:
          type: object
          allOf:
            - $ref: '#/components/schemas/SelfLink'

    # BookResultModelCollection ----------------------------------------------------
    BookResultModelCollection:
      properties:
        embedded:
          type: object
          properties:
            bookModelList:
              type: array
              items:
                type: object
                allOf:
                  - $ref: '#/components/schemas/BookResultModel'

    # Empty ------------------------------------------------------------------------
    Empty:
      type: object

    # GameModel --------------------------------------------------------------------
...
    NameModel:
      required:
        - name
      type: object
      properties:
        name:
          type: string

    # PersonModel ------------------------------------------------------------------
...
    SelfLink:
      type: object
      properties:
        self:
          allOf:
            - $ref: '#/components/schemas/UrlArray'

    # StudentModel -----------------------------------------------------------------
...
    UrlArray:
      type: array
      items:
        type: object
        properties:
          href:
            type: string
            format: url